refactor: Reorganized components to smaller sizes.
