+++ kakao-shopping/src/components/atoms/OptionDetailList.jsx
