+++ kakao-shopping/src/components/templates/MainProductsTemplate.jsx
@@ -4,7 +4,7 @@ import PageSetButton from "../molecules/PageSetButton";
 import MainProducts from "../organisms/MainProducts";
 import { useQuery } from "react-query";
 import { getProducts } from '../../apis/api';
-import Error from "../molecules/Error";
+import Error from "../atoms/Error";
 
 const MainProductsTemplate = () => {
   const storedPage = sessionStorage.getItem("currentPage");
